algorithm types
